services/query_service/pyproject.toml
[project]
name = "query-service"
version = "0.1.0"
description = "User-facing API and Orchestrator"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.123.0",
    "grpcio-tools>=1.76.0",
    "langchain>=1.1.0",
    "langchain-openai>=1.1.0",
    "rag-shared",
    "uvicorn[standard]>=0.38.0",
    "websockets>=15.0.1",
]

[tool.uv.sources]
rag-shared = { workspace = true }
